workers/semantics/navigation/002.html

<!--

/*
-->
<!doctype html>
<title>navigating 2</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
var date;
var newDate;
</script>
<iframe src="001-1.html"></iframe>
<script>
var t = async_test();
onload = t.step_func(function() {
  var iframe = document.querySelector('iframe');
  setTimeout(this.step_func(function() {
    window[0].document.links[0].click();
    iframe.onload = this.step_func(function() {
      this.onload = null;
      newDate = new Date();
      setTimeout(this.step_func(function() {
        assert_less_than(Number(date), Number(newDate));
        assert_equals(window[0].document.title, 'foo');
        this.done();
      }), 300);
    });
  }), 150);
});
</script>
<!--
*/
//-->

